Fixes wallet claim issue

+++ test/KintoID.t.sol
@@ -97,7 +97,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testMintIndividualKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         vm.startPrank(_kycProvider);
         assertEq(_kintoIDv1.isKYC(_user), false);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -111,7 +111,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testMintCompanyKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](2);
+        uint16[] memory traits = new uint16[](2);
         traits[0] = 2;
         traits[1] = 5;
         vm.startPrank(_kycProvider);
@@ -127,7 +127,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testMintIndividualKYCWithInvalidSender() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_user);
         vm.expectRevert("Invalid Provider");
@@ -136,7 +136,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testMintIndividualKYCWithInvalidSigner() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 5, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         vm.expectRevert("Invalid Signer");
@@ -145,7 +145,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testMintIndividualKYCWithInvalidNonce() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -155,7 +155,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testMintIndividualKYCWithExpiredSignature() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp - 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         vm.expectRevert("Signature has expired");
@@ -166,7 +166,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testBurnKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -178,7 +178,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testOnlyProviderCanBurnKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -199,7 +199,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
 
     function testBurningTwiceFails() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -293,7 +293,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
     function testRemoveTrait_WhenCallerIsProvider() public {
         vm.startPrank(_kycProvider);
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         _kintoIDv1.addTrait(_user, 1);
@@ -304,7 +304,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
     }
 
     function testRemoveTrait_RevertWhen_CallerIsNotProvider() public {
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         approveKYC(_kycProvider, _user, _userPk, traits);
         assertEq(_kintoIDv1.hasTrait(_user, 1), true);
@@ -325,7 +325,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
     function testProviderCanAddSanction() public {
         vm.startPrank(_kycProvider);
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         _kintoIDv1.addSanction(_user, 1);
@@ -337,7 +337,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
     function testProviderCanRemoveSancion() public {
         vm.startPrank(_kycProvider);
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         _kintoIDv1.addSanction(_user, 1);
@@ -349,7 +349,7 @@ contract KintoIDTest is KYCSignature, AATestScaffolding, UserOp {
     }
 
     function testAddSanction_RevertWhen_CallerIsNotKYCProvider() public {
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         approveKYC(_kycProvider, _user, _userPk, traits);
